Update Experiment 1
A Water Jug Problem: You are given two jugs, a 4-gallon one and a 3-gallon one, a pump which has 
unlimited water which you can use to fill the jug, and the ground on which water may be poured. Neither 
jug has any measuring markings on it. How can you get exactly 2 gallons of water in the 4-gallon jug?
Here the initial state is (0, 0). The goal state is (2, n) for any value of n.