Update Experiment 2
Missionaries and Cannibals: Missionaries and Cannibals is a problem in which 3 missionaries and 3 
cannibals want to cross from the left bank of a river to the right bank of the river. There is a boat on 
the left bank, but it only carries at most two people at a time (and can never cross with zero people). 
If cannibals ever outnumber missionaries on either bank, the cannibals will eat the missionaries.
A state can be represented by a triple, (m c b), where m is the number of missionaries on the left, c 
is the number of cannibals on the left, and b indicates whether the boat is on the left bank or right 
bank. For example, the initial state is (3 3 L) and the goal state is (3 3 R)